requirements.txt
click>=7.1.2
Flask>=1.1.2
itsdangerous>=1.1.0
Jinja2>=2.11.2
MarkupSafe>=1.1.1
Werkzeug>=1.0.1
numpy>=1.18.4
scikit-learn>=0.23.1
pickle5>=0.0.10
azure-storage-blob>=12.3.1
geopy>=1.22.0
pymongo>=3.10.1
